Functionality and Aesthetic Update
Items can now be bought from the shop and will affect the global gold
and health variables. The background for the shop was also widened, font
was changed to a more readable one and the layout was readjusted
accordingly. Shop is now removed from main level and will not appear
upon game start.