Bug Fixes
Minor update involving bug fixes. Majority of task left are parameter
passing and UI bugs. Added a dummy object to allow for the event's
sprite to remain on screen hidden above the boxes in order fo rtext to
generate properly. This will also mean that the event object can be
deleted upon game resuming so it isn't activated again.