Events and Minigames Linked
There are a lot of problems with this build. Objects need to be reviewed
to ensure that they dynamically handle like 0 health or money. Also need
to work on ensuring balance of the hunger health and money. We need to
integrate the shop event into something. And work on the object
randomisation. Them more objects there are the less of a problem this
is. The death and ending screens need reworked. Also the castle needs
fixed at the end.