Death Event and Event Objects Bug Fix and Shop Integration
Fixed a death event bug in which it would look for the value of the
button array before they were created. Also altered the event objects to
disappear when the player's health reaches 0, so as to not draw text and
boxes when they shouldn't.
Added in a signpost for shop functionality, still needs to be added into
the list of things that spawn in.