Content Creation
Added three new events to the game: sellsword, merchant and beggar.
The sellsword has a global variable associated with it initialised by
the main player object and can be used in different events where
appropriate;
The merchant acts much like the shop, drawing a pool of two items in a
very similar way.