Minor Graphics Update
Added shading/outlines to various sprites and backgrounds (changes can
be seen in the river minigame and the mountain biome). Also tweaked the
rain overlay's colour and coverage, now the sky has a gradient. To make
the sky seem darker as it rains. Finally, I noticed that the snow
overlays still had the old sky colour, this has now been rectified.