Tweak + Raft update.
Game is now half its original length (numerous playthroughs never made
it past the 30% mark originally, however a couple managed to get to the
last third.) Other than that, no balancing tweaks are needed and the
game is now possible to beat, but still challenging. Furthermore, the
player sprite is now baked into the raft sprite.

KNOWN ISSUES:
Duplicate spawning and the Wizard causing crashes. (But we all knew that
:(  )
Bartering with the Highwayman does not remove gold.
Pop-up window after the river minigame has a typo. (I tried to fix it
myself but I couldn't find where the script was kept.)

Happy hunting.