Minigame Death Fix
Fixed a previous bug which occured when the player lost enough health in
a minigame to die in the main level, I think this was caused due to too
many instances of the event parent existing at once. Now upon chosing to
play a minigame, the event that sent you there and any event parents are
destroyed.