Various Fixes and Additions
Workers Rights! There is now a constant income generated in the same
place that hunger is gradually reduced
Better Shop - The shop items now make a little more sense.  Bonus -
there is a 'mystery box' item in the shop pool which has random costs
and effects - hopefully- every time.
Re-fixed health and hunger limiting - this wasn't working quite right
before so I moved the code to do so into the lowest level of the step
event of the game controller object.